Debugging... Trying vertex buffer inputs
Vertex buffers now working

Still drawing with indexes. I am confused about the number of vertices
vs the number of indices. Shouldn't the number of indices be less? There
was a weird behavior with there being 4 duplicates of each vertex when I
would print them. I should look into that

Added uv mesh data back in

To keep consistency with the tutorial I am following. It also confirmed
my understanding of vertex attribute descriptions

Descriptor Set slight refactor

I found the creation of descriptor sets cumbersome, so I did a slight
refactoring of how they are made. I still have the
DescriptorLayoutBuilder and the DescriptorWriter, but they are now
combined in what I called the DescriptorBuilder. I realized that for
each binding you add to the layout, you have to write it too. So why not
add the binding to the writer and the layout builder at once, then when
it comes time to build the descriptor set, you build the layout first,
then write.

I also created a DescriptorSet class, which is good for portability of
descriptor set objects, but makes it hard when binding since you usually
pass an array of VkDescriptorSet handles, but you only have a list of
DescriptorSet wrapper handles, so there is an intermediate step to
create another VkDescriptorSet array.

All of this to add a uniform buffer to my mesh shader to try and debug
why the MVP matrices are not working right.

View matrices understood

The resulting render is still incorrect, there is something going on
with the perspective...

I have some shader debugging outputs in here, they will stay until I
figure out what is going on with this damn render!

BUGFIX: I think I fixed the perspective and view

I needed to swap near and far values for the perspective matrix, as well as adjust the translation
and forward direction of the view matrix. I should still try to figure
out why I need to swap them.

Okay I actually understand what is happening now..

I finally get what is happening. The near and far planes should be as
expected, not swapped. The perspective matrix derivation takes this into
account, whereas the glm version does not..