Descriptor Set slight refactor
I found the creation of descriptor sets cumbersome, so I did a slight
refactoring of how they are made. I still have the
DescriptorLayoutBuilder and the DescriptorWriter, but they are now
combined in what I called the DescriptorBuilder. I realized that for
each binding you add to the layout, you have to write it too. So why not
add the binding to the writer and the layout builder at once, then when
it comes time to build the descriptor set, you build the layout first,
then write.

I also created a DescriptorSet class, which is good for portability of
descriptor set objects, but makes it hard when binding since you usually
pass an array of VkDescriptorSet handles, but you only have a list of
DescriptorSet wrapper handles, so there is an intermediate step to
create another VkDescriptorSet array.

All of this to add a uniform buffer to my mesh shader to try and debug
why the MVP matrices are not working right.